✏ modify : store관련(뱁새, 편지, 포인트) 기능 완료

+++ Backend/store-service/src/main/java/com/zootopia/storeservice/store/entity/LetterPaperBook.java
@@ -19,7 +19,7 @@ public class LetterPaperBook {
 
     @Id
     @Column(name = "LETTER_PAPER_ID")
-    private Long letterPaperId;
+    private int letterPaperId;
 
     @Id
     @Column(name = "MEMBER_ID")
